Move LoadingScreen to common screens folder
