+++ src/ui/states/appConnect/AppConnectState.tsx
@@ -5,7 +5,7 @@ import { Text, Button, Heading, ValueCard } from 'dash-ui-kit/react'
 import { withAccessControl } from '../../components/auth/withAccessControl'
 import ButtonRow from '../../components/layout/ButtonRow'
 import { TitleBlock } from '../../components/layout/TitleBlock'
-import LoadingScreen from '../../components/layout/LoadingScreen'
+import LoadingScreen from '../../components/layout/screens/LoadingScreen'
 import { AppConnect } from '../../../types/AppConnect'
 import { AppConnectStatus } from '../../../types/enums/AppConnectStatus'
 import { getFaviconUrl } from '../../../utils'
