+++ src/ui/states/approveTransaction/ApproveTransactionState.tsx
@@ -12,7 +12,7 @@ import { useExtensionAPI, useSigningKeys } from '../../hooks'
 import { StateTransitionWASM } from 'dash-platform-sdk/types'
 import { withAccessControl } from '../../components/auth/withAccessControl'
 import type { OutletContext } from '../../types'
-import LoadingScreen from '../../components/layout/LoadingScreen'
+import LoadingScreen from '../../components/layout/screens/LoadingScreen'
 import { PublicKeySelect, type KeyRequirement } from '../../components/keys'
 import { IdentitySelect } from '../../components/identity/IdentitySelect'
 import { TransactionDetails } from './details'
